api/pyproject.toml
[tool.poetry]
name = "spc-api"
version = "0.1.0"
description = "SPC Partner Dashboard API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [ { include = "spc_api", from = "src" } ]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-cors = "^4.0.1"
flask-jwt-extended = "^4.6.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
marshmallow = "^3.21.3"
flask-marshmallow = "^1.2.1"
python-dotenv = "^1.0.1"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
marshmallow-sqlalchemy = "^1.4.2"
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest = "^8.3.2"

[tool.black]
line-length = 100
target-version = ["py311"]
include = "src|tests"

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]


